% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stages.R
\name{clean_stages}
\alias{clean_stages}
\title{Function \code{clean_stages}}
\usage{
clean_stages(stages)
}
\arguments{
\item{stages}{List of character vectors. Each character vector
is a subcollections of \code{remake} YAML files that can be run in parallel. 
The .yml extensions of the remake files are optional. For example, if I want to run 
remake1.yml and remake2.yml in parallel and then run finish.yml, I would specify
    stages = list(
      c("remake1.yml", "remake2.yml"),
      "finish.yml"
    )

or 

    stages = list(
      c("remake1", "remake2"),
      "finish"
    )}
}
\description{
Utility function to clean the \code{stages} argument to 
the function \code{remake2make}.
}
\seealso{
remake2make
}

