% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{collate_remakefiles}
\alias{collate_remakefiles}
\title{Function \code{collate_remakefiles}}
\usage{
collate_remakefiles(remakefiles)
}
\arguments{
\item{remakefiles}{names/paths of the \code{remake}/\code{YAML} files.}
}
\value{
Name of the collated (or solitary) \code{remake}/\code{YAML} file
}
\description{
Collate multiple \code{remake}/\code{YAML} files 
into a single \code{remake}/\code{YAML} file if necessary.
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\details{
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\seealso{
\code{\link{help_parallelRemake}}
}

