% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_parallelRemake}
\alias{example_parallelRemake}
\title{Function \code{example_parallelRemake}}
\usage{
example_parallelRemake(example = list_examples_parallelRemake())
}
\arguments{
\item{example}{name of the example. To see all the available example names, 
run \code{\link{list_examples_parallelRemake}}.}
}
\description{
Copy a parallelRemake example to the current working directory.
To see the names of all the examples, run \code{\link{list_examples_parallelRemake}}.
}
\seealso{
\code{\link{list_examples_parallelRemake}}, \code{\link{write_makefile}}
}

