% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stamps_files.R
\name{get_stamps_files}
\alias{get_stamps_files}
\title{Function \code{get_stamps_files}}
\usage{
get_stamps_files(files)
}
\arguments{
\item{files}{File paths (full or relative) of files to get timestamps.}
}
\description{
Get timestamps of \code{files}
}

