% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stamps_pkgs.R
\name{get_stamps_pkgs}
\alias{get_stamps_pkgs}
\title{Function \code{get_stamps_pkgs}}
\usage{
get_stamps_pkgs(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\description{
Get "Packaged" dates of R pakcages in \code{pkgs}
}

