% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{help_parallelRemake}
\alias{help_parallelRemake}
\title{Function \code{help_parallelRemake}}
\usage{
help_parallelRemake()
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\seealso{
\code{\link{write_makefile}}
}

