% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{list_examples_parallelRemake}
\alias{list_examples_parallelRemake}
\title{Function \code{list_examples_parallelRemake}}
\usage{
list_examples_parallelRemake()
}
\value{
a names of all the parallelRemake examples.
}
\description{
Return the names of all the parallelRemake examples.
}
\seealso{
\code{\link{example_parallelRemake}}, \code{\link{write_makefile}}
}

