% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{makefile_clean}
\alias{makefile_clean}
\title{Function \code{makefile_clean}}
\usage{
makefile_clean(remakefile, clean, add_args)
}
\arguments{
\item{remakefile}{name of the (collated) \code{remake}/\code{YAML} file}

\item{clean}{character vector of extra lines to run on \code{make clean}.}

\item{add_args}{extra arguments to add to \code{remake}}
}
\description{
Write the rules of the Makefile for cleaning up.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

