% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{makefile_head}
\alias{makefile_head}
\title{Function \code{makefile_head}}
\usage{
makefile_head(begin, targets)
}
\arguments{
\item{begin}{character vector of lines to prepend to the Makefile}

\item{targets}{\code{YAML}-like list of Makefile targets}
}
\description{
Print the top lines of the Makefile (prepended lines and phony targets).
}

