% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{makefile_rules}
\alias{makefile_rules}
\title{Function \code{makefile_rules}}
\usage{
makefile_rules(remakefile, targets, add_args)
}
\arguments{
\item{remakefile}{name of the (collated) \code{remake}/\code{YAML} file}

\item{targets}{\code{YAML}-like list of Makefile targets}

\item{add_args}{extra arguments to add to \code{remake}}
}
\description{
Write the rules of the Makefile for pseudo-making the targets.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

