% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_stages.R
\name{name_stages}
\alias{name_stages}
\title{Function \code{name_stages}}
\usage{
name_stages(stages)
}
\arguments{
\item{stages}{List of character vectors. Each character vector
is a subcollections of \code{remake} YAML files that can be run in parallel. 
Do not include the .yml extensions of the remake files. For example, if I want to run 
remake1.yml and remake2.yml in parallel and then run finish.yml, I would specify
    stages = list(
      c("remake1", "remake2"),
      "finish"
    )}
}
\description{
Utility function to assign names the \code{stages} argument to 
the function \code{remake2make}.
}
\seealso{
remake2make
}

