% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{parallelRemake_help}
\alias{parallelRemake_help}
\title{Function \code{parallelRemake_help}}
\usage{
parallelRemake_help()
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\seealso{
\code{\link{write_makefile}}
}

