% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_command.R
\name{parse_command}
\alias{parse_command}
\title{Function \code{parse_command}}
\usage{
parse_command(str)
}
\arguments{
\item{str}{Charater containing the command.}
}
\value{
List of information about the command.
}
\description{
Parse a `remake` command. 
COPIED FROM https://github.com/richfitz/remake parse_command.R.
ALL CREDIT FOR THIS FILE GOES TO RICH FITZJOHN (https://richfitz.github.io/)
}

