% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{recallable}
\alias{recallable}
\title{Function \code{recallable}}
\usage{
recallable(cache = file.path(".remake", "objects"))
}
\arguments{
\item{cache}{Character vector, path to \code{storr} cache to load from.}
}
\value{
A character vector of objects/targets available to laod from the cache.
}
\description{
Lists the objects/targets available to laod from the cache.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

