% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_utils.R
\name{remake_args}
\alias{remake_args}
\title{Function \code{remake_args}}
\usage{
remake_args(x)
}
\arguments{
\item{x}{List of arguments to remake. Must be named.}
}
\description{
Check additional arguments to \code{remake}.
}

