% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remember.R
\name{remember}
\alias{remember}
\title{Function \code{remember}}
\usage{
remember(..., cache = ".remake/objects")
}
\arguments{
\item{...}{Objects to load from cache}

\item{cache}{Character vector, path to \code{storr} cache to load from.}
}
\value{
An object from the cache
}
\description{
Loads an object from the \code{storr} cache. Defaults
to the cache used by \code{remake}.
}

