% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_example}
\alias{run_example}
\title{Function \code{run_example}}
\usage{
run_example()
}
\description{
Run the whole example: generate the \code{remake} file,
then generate the master \code{Makefile} to run the targets in parallel,
and then run \code{make} to run the workflow in parallel.
}

