% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_example_parallelRemake}
\alias{run_example_parallelRemake}
\title{Function \code{run_example_parallelRemake}}
\usage{
run_example_parallelRemake()
}
\description{
Runs the whole package example from start to finish.
Generates the files \code{code.R} and \code{remake.yml}, produces the 
Makefile, and then runs \code{make} to execute the workflow in parallel.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

