% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_if_outdated.R
\name{run_if_outdated}
\alias{run_if_outdated}
\title{Function \code{run_if_outdated}}
\usage{
run_if_outdated(expr, dep = NULL, out = NULL)
}
\arguments{
\item{expr}{R expression with code to be evaluated.}

\item{dep}{Character vector. Elements are either paths to dependency
files or R packages. Dependency files get precedence, meaning that
the package \code{MASS} will checked only if the file \code{MASS} does 
not exist.}

\item{out}{Character vector of relative (or full) paths to output files saved
by the R code in \code{expr}.}
}
\description{
Runs R code in \code{expr} if any output files \code{out}
were created before any of the dependency files \code{dep} were created
or R packages \code{pkg} were last updated.
}
\details{
. If \code{out} is empty, \code{expr} does not run. If
\code{dep} is empty, then \code{expr} runs if and only if at least
one file given in \code{out} does not exist. If both \code{dep} and
\code{out} are nonempty and every element of \code{dep} is either
an existing file or an installed R package, \code{expr} will run 
if and only if at least one dependency was modified more recently 
than at least one output file.
}

