% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{setup_example_parallelRemake}
\alias{setup_example_parallelRemake}
\title{Function \code{setup_example_parallelRemake}}
\usage{
setup_example_parallelRemake()
}
\description{
Writes the files \code{code.R} and \code{remake.yml} for 
the package example and then produces the Makefile.
}

