% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_field.R
\name{single_field}
\alias{single_field}
\title{Function \code{step}}
\usage{
single_field(field, name = "", level = 0)
}
\arguments{
\item{field}{a field or list of fields}

\item{name}{the name of the field}

\item{level}{level of recursion}
}
\description{
Recursive function used by \code{step} to turn a list of fields 
into a \code{remake} YAML file.
}
\seealso{
\code{one_process}
}

