% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{strings}
\alias{strings}
\title{Function \code{strings}}
\usage{
strings(...)
}
\arguments{
\item{...}{list of expressions to turn into character strings}
}
\value{
a named character vector
}
\description{
Turns expressions into character strings
}

