% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{write_example_parallelRemake}
\alias{write_example_parallelRemake}
\title{Function \code{write_example_parallelRemake}}
\usage{
write_example_parallelRemake()
}
\description{
Writes the files \code{code.R} and \code{remake.yml} 
for the package example.
Use the \code{\link{parallelRemake_help}} function to get more help.
}
\details{
Use the \code{\link{parallelRemake_help}} function to get more help.
}
\seealso{
\code{\link{parallelRemake_help}}
}

