% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{Function \code{write_makefile}}
\usage{
write_makefile(makefile = "Makefile", remakefiles = "remake.yml",
  begin = NULL, clean = NULL, remake_args = list())
}
\arguments{
\item{makefile}{name of Makefile to write.}

\item{remakefiles}{Character vector of paths to input \code{remake} files.}

\item{begin}{Character vector of lines to prepend to the Makefile.}

\item{clean}{Character vector of commands to add to the \code{clean} rule.}

\item{remake_args}{Fully-named list of additional arguments to \code{remake::make}.}
}
\description{
Write a master Makefile to run \code{remake} targets in parallel.
}

