% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_step.R
\name{write_step}
\alias{write_step}
\title{Function \code{write_step}}
\usage{
write_step(fields, file = "remake.yml")
}
\arguments{
\item{fields}{a nested list defnining the fields of the \code{remake} YAML file}

\item{file}{Path to output remakefile.}
}
\description{
Creates a \code{remake} YAML file (github.com/richfitz/remake) to define 
a sequential step within a parallelized stage of a workflow. See the README at 
github.com/wlandau/workflowHelper for details.
}

