% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_workflow.R
\name{write_workflow}
\alias{write_workflow}
\title{Function \code{write_workflow}}
\usage{
write_workflow(stages, file = "Makefile")
}
\arguments{
\item{file}{Path to output Makefile.}
}
\description{
Takes a collection of \code{remake} YAML files (github.com/richfitz/remake)
and creates an overarching Makefile to run them all in parallelizable stages 
(with \code{make -j} on the command line).
}

