% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_utils.R
\name{yaml_read}
\alias{yaml_read}
\title{Function \code{yaml_read}}
\usage{
yaml_read(filename)
}
\arguments{
\item{filename}{Character, name of \code{YAML} file to read}
}
\description{
COPIED FROM https://github.com/richfitz/remake utils.R.
ALL CREDIT FOR THIS FILE GOES TO RICH FITZJOHN (https://richfitz.github.io/).
}
\details{
LICENSE: https://raw.githubusercontent.com/richfitz/remake/master/LICENSE

The MIT License (MIT)

Copyright (c) <2016> <Richard G. FitzJohn>

Permission is hereby granted, free of charge, to any person 
obtaining a copy of this software and associated documentation 
files (the "Software"), to deal in the Software without 
restriction, including without limitation the rights to use, 
copy, modify, merge, publish, distribute, sublicense, and/or
sell copies of the Software, and to permit persons to whom the 
Software is furnished to do so, subject to the following 
conditions:

The above copyright notice and this permission notice shall be
 included in all copies or substantial portions of the
 Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, 
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES 
OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND 
NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT 
HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, 
WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR 
OTHER DEALINGS IN THE SOFTWARE.
}

