% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{clean_example_parallelRemake}
\alias{clean_example_parallelRemake}
\title{Function \code{clean_example_parallelRemake}}
\usage{
clean_example_parallelRemake()
}
\description{
Cleans up files generated by the package example.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

