% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_yaml.R
\name{collate_yaml}
\alias{collate_yaml}
\title{Function \code{collate_yaml}}
\usage{
collate_yaml(remakefiles = "remake.yml", collated = paste0("collated_",
  remakefiles[1]))
}
\arguments{
\item{remakefiles}{Character vector of the paths to the remake/YAML
files to collate.}

\item{collated}{name of collated remake/YAML file.}
}
\description{
Collate interdependent remake/YAML files into a single 
master remake/YAML file. Files in the "include:" fields will be collated
as well.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

