% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_utils.R
\name{include_yaml}
\alias{include_yaml}
\title{Function \code{include_yaml}}
\usage{
include_yaml(remakefiles, out = NULL)
}
\arguments{
\item{remakefiles}{Character vector of remake files to add to the collation.}

\item{out}{Character vector of names of remakefiles to collate so far.}
}
\description{
Recursively find YAML files
}

