% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{makefile}
\alias{makefile}
\title{Function \code{makefile}}
\usage{
makefile(targets = "all", remakefiles = "remake.yml", prepend = NULL,
  clean = NULL, remake_args = list(), run = TRUE, command = "make")
}
\arguments{
\item{targets}{character vector of targets to make}

\item{remakefiles}{Character vector of paths to input \code{remake} files.}

\item{prepend}{Character vector of lines to prepend to the Makefile.}

\item{clean}{Character vector of commands to add to the \code{clean} rule.}

\item{remake_args}{Named list of additional arguments to \code{remake::make}.}

\item{run}{logical, whether to actually run the Makefile or just write it.}

\item{command}{character scalar, command to run to execute the Makefile
You cannot set \code{target_names} or \code{remake_file} this way because 
those names are already reserved.}
}
\description{
Writes an executes a Makefile to distribute \code{remake} targets
over simultaneous processes.
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\details{
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\seealso{
\code{\link{help_parallelRemake}}
}

