% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{recall}
\alias{recall}
\title{Function \code{recall}}
\usage{
recall(..., cache = ".remake/objects")
}
\arguments{
\item{...}{Objects to load from cache}

\item{cache}{Character vector, path to \code{storr} cache to load from.}
}
\value{
A loaded object
}
\description{
Load an intermediate object previously generated in the workflow.
This could be a dataset, analysis, etc.
}

