% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{recall}
\alias{recall}
\title{Function \code{recall}}
\usage{
recall(..., cache = file.path(".remake", "objects"))
}
\arguments{
\item{...}{Characters, names of objects to load from cache}

\item{cache}{Character vector, path to \code{storr} cache to load from.}
}
\value{
A loaded object
}
\description{
Loads an object/target from the cache.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

