% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{recallable}
\alias{recallable}
\title{Function \code{recallable}}
\usage{
recallable(cache = file.path(".remake", "objects"))
}
\arguments{
\item{cache}{Character vector, path to \code{storr} cache to load from.}
}
\value{
A character vector of objects/targets available to laod from the cache.
}
\description{
Lists the objects/targets available to laod from the cache.
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\details{
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\seealso{
\code{\link{help_parallelRemake}}
}

