% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{recallable}
\alias{recallable}
\title{Function \code{recallable}}
\usage{
recallable(cache = file.path(".remake", "objects"))
}
\arguments{
\item{cache}{Character vector, path to \code{storr} cache to load from.}
}
\value{
A loaded object
}
\description{
List the items available to laod from the cache.
}

