% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_example_parallelRemake}
\alias{run_example_parallelRemake}
\title{Function \code{run_example_parallelRemake}}
\usage{
run_example_parallelRemake(intern = F)
}
\arguments{
\item{intern}{Argument to system, option to capture output.}
}
\description{
Run the whole example: generate the \code{remake} file,
then generate the master \code{Makefile} to run the targets in parallel,
and then run \code{make} to run the workflow in parallel.
}

