% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{setup_example_parallelRemake}
\alias{setup_example_parallelRemake}
\title{Function \code{setup_example}}
\usage{
setup_example_parallelRemake()
}
\description{
Write the files \code{code.R} and \code{remake.yml} for the example. 
and then write the Makefile.
}

