% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{write_example_parallelRemake}
\alias{write_example_parallelRemake}
\title{Function \code{write_example_parallelRemake}}
\usage{
write_example_parallelRemake()
}
\description{
Writes the files \code{code.R} and \code{remake.yml} 
for the package example.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

