% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{DEPRECATED function \code{write_makefile}}
\usage{
write_makefile(targets = "all", remakefiles = "remake.yml",
  prepend = NULL, remake_args = list(verbose = TRUE), run = TRUE,
  command = "make", args = character(0))
}
\arguments{
\item{targets}{See \code{\link{makefile}()}.}

\item{remakefiles}{See \code{\link{makefile}()}.}

\item{prepend}{See \code{\link{makefile}()}.}

\item{remake_args}{See \code{\link{makefile}()}.}

\item{run}{See \code{\link{makefile}()}.}

\item{command}{See \code{\link{makefile}()}.}

\item{args}{See \code{\link{makefile}()}.}
}
\description{
See \code{\link{makefile}}. Makefiles are not standalone anymore.
You have to write and execute them in one step with \code{link{makefile}(..., run = TRUE)}.
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\details{
Use the \code{\link{help_parallelRemake}} function to get more help.
}
\seealso{
\code{\link{makefile}}, \code{\link{help_parallelRemake}}
}

