% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{Function \code{write_makefile}}
\usage{
write_makefile(makefile = "Makefile", remakefiles = "remake.yml",
  begin = NULL, clean = NULL, remake_args = list())
}
\arguments{
\item{makefile}{Name of the Makefile to write.}

\item{remakefiles}{Character vector of paths to input \code{remake} files.}

\item{begin}{Character vector of lines to prepend to the Makefile.}

\item{clean}{Character vector of commands to add to the \code{clean} rule.}

\item{remake_args}{Named list of additional arguments to \code{remake::make}.
You cannot set \code{target_names} or \code{remake_file} this way because 
those names are already reserved.}
}
\description{
Writes a master Makefile that can distribute \code{remake} workflows
over simultaneous processes.
Read https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd for more.
}
\details{
See https://github.com/wlandau/parallelRemake/blob/master/vignettes/parallelRemake.Rmd
for a package tutorial.
}

