% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{Function \code{write_makefile}}
\usage{
write_makefile(makefile = "Makefile", remakefile = "remake.yml",
  begin = NULL, clean = NULL)
}
\arguments{
\item{makefile}{name of Makefile to write.}

\item{remakefile}{name of input \code{remake} file.}

\item{begin}{Character vector of lines to prepend to the Makefile.}

\item{clean}{Character vector of commands to add to the \code{clean} rule.}
}
\description{
Write a master Makefile to run \code{remake} targets in parallel.
}

