% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{Function \code{write_makefile}}
\usage{
write_makefile(stages, file = "Makefile", clean = NULL)
}
\arguments{
\item{stages}{List of character vectors. Organizes the steps of a 
workflow into parallelizable stages. Each character string is a 
path to a \code{remake}/YAML file encoding a step in the workflow.
The ".yml" file extension is optional.}

\item{file}{Path to output Makefile.}

\item{clean}{Character vector of commands to add to the \code{clean} rule.}
}
\description{
Takes a collection of \code{remake} YAML files (github.com/richfitz/remake)
and creates an overarching Makefile to run them all in parallelizable stages 
(with \code{make -j} on the command line).
}

