% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_workflow.R
\name{write_workflow}
\alias{write_workflow}
\title{Function \code{write_workflow}}
\usage{
write_workflow(stages, file = "Makefile")
}
\arguments{
\item{stages}{List of character vectors. Each character vector
is a subcollections of \code{remake} YAML files that can be run in parallel. 
Do not include the .yml extensions of the remake files. For example, if I want to run 
remake1.yml and remake2.yml in parallel and then run finish.yml, I would specify
    stages = list(
      c("remake1", "remake2"),
      "finish"
    )}

\item{file}{Path to output Makefile.}
}
\description{
Takes a collection of \code{remake} YAML files (github.com/richfitz/remake)
and creates a Makefile to run them all.
}
\details{
A disadvantage of remake (github.com/richfitz/remake) is that recipes cannot
be executed in parallel. This function provides one possible solution: namely, to 
break up the workflow into multiple remake YAML files and write an overarching
Makefile to execute subcollections of recipes in parallel. For example, suppose I want a 
want to run remake1.yml, remake2.yml, and remake3.yml in parallel and then run
remake4.yml, remake5.yml, and remake6.yml in parallel. I can create a Makefile to do this 
by first defining a list of stages. 
    stages = list(
      stage1 = paste0("remake", 1:3),
      stage2 = paste0("remake", 4:6)
    )
Then, I call makefile(stages) to produce the approprate Makefile. 
Lastly, I call \code{make -j 3} to run the Makefile to run 
the \code{remake} files in each stage over 3 separate parallel processes.
}

