% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_utils.R
\name{yaml_read}
\alias{yaml_read}
\title{Function \code{yaml_read}}
\usage{
yaml_read(filename)
}
\arguments{
\item{filename}{Character, name of \code{YAML} file to read}
}
\description{
COPIED FROM https://github.com/richfitz/remake utils.R.
ALL CREDIT FOR THIS FILE GOES TO RICH FITZJOHN (https://richfitz.github.io/).
}

