% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_utils.R
\name{yaml_yesno_truefalse}
\alias{yaml_yesno_truefalse}
\title{Function \code{yaml_yesno_truefalse}}
\usage{
yaml_yesno_truefalse(file)
}
\arguments{
\item{file}{Character, name of \code{YAML} file to read}
}
\description{
Turn yes/no values of a YAML file into TRUE/FALSE
}

